% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARGEE-package.R
\docType{package}
\name{MARGEE-package}
\alias{MARGEE-package}
\title{Mixed-model Association robust single-variant tests of Gene–Environment interactions for common variants}
\description{
Perform generalized linear mixed robust single-variant main effect tests, gene-environment interaction tests for common variants
}
\author{
Mengyu Zhang \email{mengyu.zhang@uth.tmc.edu}
}
