% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rom.R
\name{rom}
\alias{rom}
\title{Conduct robust single-variant -environment interaction tests and joint test for common variants}
\usage{
rom(
  null.obj,
  outdir,
  interaction,
  geno.file,
  interaction.covariates = NULL,
  meta.output = F,
  center = T,
  related.id,
  MAF.range = c(0.01, 0.5),
  miss.cutoff = 1,
  missing.method = "impute2mean",
  nperbatch = 100,
  ncores = 1,
  is.dosage = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{null.obj}{The 'glmmkin' class object. Please use sparse kinship matrix (e.g. dsTMatrix class) as kin matrix when use 'glmmkin' function in GMMAT package}

\item{outdir}{the directory name of outfile res.txt}

\item{interaction}{a vector contains indices or variable name of the environmental factors}

\item{geno.file}{the name of a GDS file (including the suffix .gds)}

\item{interaction.covariates}{a vector contains indices or variable name of the interaction covariates}

\item{meta.output}{boolean value to modiy the output file. If TRUE, the GxE effect estimate and variance and covariance associated with the effect estimate are included in the output file. (default = FALSE)}

\item{center}{If TRUE, genotypes will be centered before tests. Otherwise, original values will be used in the tests (default = TRUE)}

\item{related.id}{a numeric or a character vector contains family id}

\item{MAF.range}{a numeric vector of length 2 defining the minimum and maximum minor allele frequencies of variants that should be included in the analysis (default = c(0.01, 0.5), i.e. common variants).}

\item{miss.cutoff}{the maximum missing rate allowed for a variant to be included (default = 1, including all variants)}

\item{missing.method}{method of handling missing genotypes.Either "impute2mean" or "omit" (default = "impute2mean") \link{TBT}}

\item{nperbatch}{an positive integer for how many SNPs should be tested in a batch (default = 100).}

\item{ncores}{a positive integer indicating the number of cores used in parallel computing (default = 1)}

\item{is.dosage}{whether imputed dosage should be used from a GDS infile (default = FALSE).}

\item{verbose}{whether failed matrix inversions should be written to outfile.err for debugging (default = FALSE).}
}
\description{
Conduct robust single-variant -environment interaction tests and joint test for common variants
}
