% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoM-package.R
\docType{_PACKAGE}
\name{RoM-package}
\alias{RoM-package}
\title{Robust Mixed-model Association Test for Gene-Environment Interaction}
\description{
Perform single variant robust gene-environment interaction test and joint test for common variants
}
\author{
Mengyu Zhang \email{mengyu.zhang@uth.tmc.edu}
}
